requirements.txt
easygui==0.98.1
numpy==1.18.2
opencv_python==4.2.0.34
Pillow==7.2.0
